refactor: extracing various formatting functions, round CPU hours in graph
